automobile-dev-new/requirements.txt
astroid==1.6.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.6
django-cas-ng==3.5.9
django-cors-headers==2.4.0
django-extensions==2.0.7
django-simple-sso==0.13.2
djangorestframework==3.8.2
dotplus==0.3
graphviz==0.8.4
gunicorn==19.8.1
idna==2.6
isort==4.3.4
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
jsonpickle==0.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==4.1.0
mysqlclient==1.3.12
numpy==1.15.1
passlib==1.7.1
pbr==4.0.2
pluggy==0.6.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.5.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pylint==1.9.2
pytest==3.5.1
python-cas==1.2.0
pytz==2018.5
requests==2.18.4
simplejson==3.16.0
six==1.11.0
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.22
webservices==0.7
whitenoise==3.3.1
wrapt==1.10.11
